python/requirements.txt
RPi.GPIO
flask
